DDCNL-8950: Fixed circular dependency involving TaxAccountSummaryService and TotalTaxService
