+++ app/uk/gov/hmrc/tai/service/TaxAccountSummaryService.scala
@@ -20,22 +20,22 @@ import com.google.inject.{Inject, Singleton}
 import play.api.mvc.Request
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.tai.connectors.TaxAccountConnector
 import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent
 import uk.gov.hmrc.tai.model.domain.formatters.TaxAccountSummaryHodFormatters
 import uk.gov.hmrc.tai.model.domain._
 import uk.gov.hmrc.tai.model.domain.formatters.taxComponents.TaxAccountHodFormatters
 import uk.gov.hmrc.tai.model.tai.TaxYear
+import uk.gov.hmrc.tai.service.helper.TaxAccountHelper
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.language.postfixOps
 
 @Singleton
 class TaxAccountSummaryService @Inject() (
-  taxAccountConnector: TaxAccountConnector,
   codingComponentService: CodingComponentService,
   incomeService: IncomeService,
-  totalTaxService: TotalTaxService
+  totalTaxService: TotalTaxService,
+  taxAccountHelper: TaxAccountHelper
 )(implicit
   ec: ExecutionContext
 ) extends TaxAccountSummaryHodFormatters with TaxAccountHodFormatters {
@@ -45,7 +45,7 @@ class TaxAccountSummaryService @Inject() (
     request: Request[_]
   ): Future[TaxAccountSummary] =
     for {
-      totalEstimatedTax       <- totalEstimatedTax(nino, year)
+      totalEstimatedTax       <- taxAccountHelper.totalEstimatedTax(nino, year)
       taxFreeAmountComponents <- codingComponentService.codingComponents(nino, year)
       taxCodeIncomes          <- incomeService.taxCodeIncomes(nino, year)
       totalTax                <- totalTaxService.totalTax(nino, year)
@@ -71,21 +71,6 @@ class TaxAccountSummaryService @Inject() (
         taxFreeAllowance
       )
     }
-
-  def totalEstimatedTax(nino: Nino, year: TaxYear)(implicit hc: HeaderCarrier): Future[BigDecimal] = {
-    val componentTypesCanAffectTotalEst: Seq[TaxComponentType] =
-      Seq(UnderPaymentFromPreviousYear, OutstandingDebt, EstimatedTaxYouOweThisYear)
-
-    taxAccountConnector
-      .taxAccount(nino, year)
-      .flatMap { taxAccount =>
-        val totalTax = taxAccount.as[BigDecimal](taxAccountSummaryReads)
-        val componentsCanAffectTotal = taxAccount
-          .as[Seq[CodingComponent]](codingComponentReads)
-          .filter(c => componentTypesCanAffectTotalEst.contains(c.componentType))
-        Future(totalTax + componentsCanAffectTotal.map(_.inputAmount.getOrElse(BigDecimal(0))).sum)
-      }
-  }
 
   private[service] def taxFreeAmountCalculation(codingComponents: Seq[CodingComponent]): BigDecimal =
     codingComponents.foldLeft(BigDecimal(0))((total: BigDecimal, component: CodingComponent) =>
